moved mock platform, removed vjs dependency
