+++ lib/platform/index.js
@@ -1,6 +1,4 @@
 'use strict'
-var Observable = require('vigour-js/lib/observable')
-
 var agent = require('vigour-ua')(navigator.userAgent)
 var env = global.env
 var target = env && env.target
@@ -14,5 +12,5 @@ if (target === 'android' || target === 'ios') {
 } else if (browser && browser === 'chrome') {
   module.exports = require('./web')
 } else {
-  module.exports = new Observable(false)
+  module.exports = require('./mock')
 }
